+++ scripts/resolve.vapor_model.php
@@ -21,7 +21,7 @@
 /**
  * @var xPDOTransport $transport
  * @var array $object
- * @var array $options
+ * @var array $fileMeta
  */
 $transport->xpdo->loadClass('vapor.vaporVehicle', MODX_CORE_PATH . 'components/vapor/model/', true, true);
 return true;