+++ changelog.md
@@ -1,5 +1,8 @@
 # MODX Vapor Changelog
 
+- [#8817](http://tracker.modx.com/issues/8817) Fix media source import via resolvers
+- Fix invalid reference from $options to $fileMeta in resolvers
+- [#8871](http://tracker.modx.com/issues/8871) Fix import of extra tables without a table_prefix
 - Add excludeExtraTablePrefix to avoid prepending table_prefix for specific tables
 - Add excludeExtraTables option for excluding specific extra non-core tables
 - Add excludeFiles option for excluding specific files/dirs from base_path